refactored contact
